refactor: changes graham suggested in pr review
