+++ .storybook/config.js
@@ -3,18 +3,23 @@ import { addDecorator } from '@storybook/react';
 import { checkA11y } from '@storybook/addon-a11y';
 import { withConsole } from '@storybook/addon-console';
 
-import 'typeface-pt-serif/index.css'
-import 'typeface-pt-sans-caption/index.css'
-import { StyleDecorator } from './StyleDecorator';
+import 'typeface-pt-serif/index.css';
+import 'typeface-pt-sans-caption/index.css';
+import { fontDecorator } from './StyleDecorator';
 
-addDecorator((storyFn, context) => withConsole()(storyFn)(context))
-addDecorator(checkA11y)
-addDecorator(StyleDecorator);
+const fontStyle = {
+  fontFamily: 'PT Sans Caption'
+};
+const fontDecorator = storyFn => <div style={fontStyle}>{storyFn()}</div>;
+
+addDecorator((storyFn, context) => withConsole()(storyFn)(context));
+addDecorator(checkA11y);
+addDecorator(fontDecorator);
 
 function loadStories() {
-  const req = require.context('../src', true, /\.storybook\.tsx$/)
+  const req = require.context('../src', true, /\.storybook\.tsx$/);
 
-  req.keys().forEach(file => req(file))
+  req.keys().forEach(file => req(file));
   // You can require as many stories as you need.
 }
