--- src/components/MobileNavBar/NavBar.storybook.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import { RouterDecorator } from '../../RouterDecorator';
-import { MobileNavBar } from '.';
-
-storiesOf('Mobile NavBar', module)
-  .addDecorator(RouterDecorator)
-  .add('default render', () => <MobileNavBar />);
