+++ src/components/NavBar/index.tsx
@@ -16,7 +16,7 @@ export interface NavBarProps {
 }
 
 export const HeaderLogo: React.SFC = () => (
-  <a href={'https://www.smarterbalanced.org/'}>
+  <a href="https://www.smarterbalanced.org/">
     <img alt="Smarter Balanced logo" src={sbLogo} className="sb-logo" />
     <style jsx>{`
       .sb-logo {
